Separate module for cmd, moved commands to cld
