+++ tools/proposal/timelock/mcm_with_timelock_test.go
@@ -2,8 +2,10 @@ package timelock
 
 import (
 	"crypto/ecdsa"
+	"encoding/json"
 	"errors"
 	"math/big"
+	"os"
 	"testing"
 	"time"
 
@@ -1380,4 +1382,33 @@ func TestE2E_ValidBypassProposalOneBatchTx(t *testing.T) {
 	hasRole, err = timelock.HasRole(&bind.CallOpts{}, cancellerRole, auths[0].From)
 	assert.NoError(t, err)
 	assert.True(t, hasRole)
+}
+
+func TestTimelockProposalFromFile(t *testing.T) {
+	mcmsProposal := MCMSWithTimelockProposal{
+		MCMSProposal: mcms.MCMSProposal{
+			Version:              "1",
+			ValidUntil:           100,
+			Signatures:           []mcms.Signature{},
+			OverridePreviousRoot: false,
+			Description:          "Test Proposal",
+			ChainMetadata:        make(map[mcms.ChainIdentifier]mcms.ChainMetadata),
+		},
+		TimelockAddresses: make(map[mcms.ChainIdentifier]common.Address),
+		Transactions:      make([]BatchChainOperation, 0),
+		Operation:         Schedule,
+		MinDelay:          "1h",
+	}
+
+	tempFile, err := os.CreateTemp("", "timelock.json")
+	assert.NoError(t, err)
+
+	proposalBytes, err := json.Marshal(mcmsProposal)
+	assert.NoError(t, err)
+	err = os.WriteFile(tempFile.Name(), proposalBytes, 0644)
+	assert.NoError(t, err)
+
+	fileProposal, err := NewMCMSWithTimelockProposalFromFile(tempFile.Name())
+	assert.NoError(t, err)
+	assert.Equal(t, mcmsProposal, *fileProposal)
 }
