+++ tools/proposal/types.go
@@ -1,6 +1,11 @@
 package proposal
 
-import "github.com/smartcontractkit/ccip-owner-contracts/tools/proposal/mcms"
+import (
+	"errors"
+
+	"github.com/smartcontractkit/ccip-owner-contracts/tools/proposal/mcms"
+	"github.com/smartcontractkit/ccip-owner-contracts/tools/proposal/timelock"
+)
 
 type ProposalType string
 
@@ -20,4 +25,15 @@ type Proposal interface {
 	ToExecutor(sim bool) (*mcms.Executor, error)
 	AddSignature(signature mcms.Signature)
 	Validate() error
+}
+
+func LoadProposal(proposalType ProposalType, filePath string) (Proposal, error) {
+	switch proposalType {
+	case MCMS:
+		return mcms.NewProposalFromFile(filePath)
+	case MCMSWithTimelock:
+		return timelock.NewMCMSWithTimelockProposalFromFile(filePath)
+	default:
+		return nil, errors.New("unknown proposal type")
+	}
 }
